third_party/blink/web_tests/fast/dom/shadow/layout-tests-can-access-shadow-expected.txt

This tests that layout tests can access shadow DOM.


PASS shadow.nodeName is "#document-fragment"
PASS internals.shadowRoot(shadow) threw exception TypeError: Failed to execute 'shadowRoot' on 'Internals': parameter 1 is not of type 'Element'..
PASS internals.shadowRoot(p) is null
PASS shadow.nodeName is "#document-fragment"
PASS shadow === internals.shadowRoot(p) is true
PASS successfullyParsed is true

TEST COMPLETE
